+++ lib/concierge/context/null.rb
@@ -5,9 +5,11 @@ class Concierge::Context
   # This is a null object implementing the contract expected by +Concierge::Context+.
   # It responds to the following methods:
   #
-  # +events+  - always returns am empty list of events.
-  # +augment+ - a no-op, returning +nil+.
-  # +to_h+    - returns an empty Hash.
+  # +events+   - always returns am empty list of events.
+  # +augment+  - a no-op, returning +nil+.
+  # +enable!+  - no-op
+  # +disable!+ - no-op
+  # +to_h+     - returns an empty Hash.
   #
   # By default, on applications that do not need context tracking, the context
   # should be set to an instance of this class.
@@ -19,6 +21,12 @@ class Concierge::Context
     end
 
     def augment(event)
+    end
+
+    def enable!
+    end
+
+    def disable!
     end
 
     def to_h